requirements.txt
pandas==2.1.4
configparser==6.0.0
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9